backend/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
python-multipart==0.0.20
torch>=2.0.0
torchvision>=0.15.0
pennylane==0.37.0
autoray==0.6.12
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
opencv-python-headless>=4.8.0
